pyproject.toml
[project]
name = "capstone2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "dash>=3.4.0",
    "dash-bootstrap-components>=2.0.4",
    "duckdb>=1.4.3",
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "jupyter>=1.1.1",
    "kaleido>=1.2.0",
    "lightgbm>=4.6.0",
    "mangum>=0.20.0",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "plotly>=6.5.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "uvicorn>=0.40.0",
    "xgboost>=3.1.3",
]
